refactor: extract panel keywords to commandutil
