+++ src/main/java/seedu/smartnus/logic/parser/ParserUtil.java
@@ -3,9 +3,9 @@ package seedu.smartnus.logic.parser;
 import static java.util.Objects.requireNonNull;
 import static seedu.smartnus.commons.core.Messages.MESSAGE_INVALID_LIMIT_ARG;
 import static seedu.smartnus.commons.util.CollectionUtil.requireAllNonNull;
-import static seedu.smartnus.logic.commands.ListCommand.NOTE_KEYWORD;
-import static seedu.smartnus.logic.commands.ListCommand.QUESTION_KEYWORD;
-import static seedu.smartnus.logic.commands.ListCommand.TAG_KEYWORD;
+import static seedu.smartnus.logic.commands.CommandUtil.NOTE_KEYWORD;
+import static seedu.smartnus.logic.commands.CommandUtil.QUESTION_KEYWORD;
+import static seedu.smartnus.logic.commands.CommandUtil.TAG_KEYWORD;
 import static seedu.smartnus.logic.commands.ThemeCommand.DARK_KEYWORD;
 import static seedu.smartnus.logic.commands.ThemeCommand.LIGHT_KEYWORD;
 import static seedu.smartnus.logic.parser.AddTfqCommandParser.ANSWER_FALSE;
